+++ test/updox/models/location_test.rb
@@ -34,8 +34,6 @@ class LocationTest < Minitest::Test
         end
 
         it 'calls sync api' do
-          request_body = JSON.parse(@request.body)
-
           assert_equal(1, request_body['locations'].size)
           assert_equal(loc.to_h, request_body['locations'].first)
         end
@@ -43,14 +41,12 @@ class LocationTest < Minitest::Test
         it 'has bulk call' do
           loc.class.sync([loc, loc], account_id: account_id)
 
-          request_body = JSON.parse(@request.body)
-
           assert_equal(2, request_body['locations'].size)
           assert_equal(loc.to_h, request_body['locations'].first)
         end
 
         it 'has acct auth' do
-          assert_acct_auth(@request, account_id)
+          assert_acct_auth(account_id)
         end
       end
 
@@ -109,7 +105,7 @@ class LocationTest < Minitest::Test
         end
 
         it 'has acct auth' do
-          assert_acct_auth(@request, account_id)
+          assert_acct_auth(account_id)
         end
       end
     end